requirements.txt
django==3.2
djangorestframework==3.13.1
djoser==2.1.0